+++ frontend/where_child_bus/lib/models/create_nursery_error.dart
@@ -1,0 +1,8 @@
+enum CreateNurseryError {
+  unknown,
+  invalidEmail,
+  emailAlreadyExist,
+  networkError,
+  passwordsDoNotMatch,
+  fieldsDoNotFilled,
+}
